services/gateway/pyproject.toml
[project]
name = "agent-platform-gateway"
version = "0.1.0"
description = "GraphQL API Gateway for the Agent Platform"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "strawberry-graphql[fastapi]>=0.230",
  "uvicorn[standard]>=0.30",
  "httpx>=0.27",
  "pydantic>=2.8",
  "structlog>=24.0",
  "PyJWT>=2.8",
  "agent-platform-common",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.24",
  "httpx",
]

[tool.uv.sources]
agent-platform-common = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gateway"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
